# internal/server/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Basic authentication for OAuth2 client credentials",
        "title": "Easyflow OAuth2 Server API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "email": "support@easyflow.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/admin/stats": {
            "get": {
                "description": "Retrieve system statistics including user, client, and session counts",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin"
                ],
                "summary": "Get system statistics",
                "responses": {
                    "200": {
                        "description": "System statistics",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Failed to retrieve stats",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/admin/system-info": {
            "get": {
                "description": "Retrieve system information including version and health status",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin"
                ],
                "summary": "Get system information",
                "responses": {
                    "200": {
                        "description": "System information",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Failed to retrieve system info",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/auth/login": {
            "post": {
                "description": "Authenticate a user and create a session",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Authentication"
                ],
                "summary": "User login",
                "parameters": [
                    {
                        "description": "Login credentials",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_server_routes_auth.LoginRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Login successful, session token set in cookie",
                        "schema": {
                            "$ref": "#/definitions/internal_server_routes_auth.LoginResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request payload",
                        "schema": {
                            "$ref": "#/definitions/easyflow-oauth2-server_internal_shared_errors.APIError"
                        }
                    },
                    "401": {
                        "description": "Invalid credentials",
                        "schema": {
                            "$ref": "#/definitions/easyflow-oauth2-server_internal_shared_errors.APIError"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/easyflow-oauth2-server_internal_shared_errors.APIError"
                        }
                    }
                }
            }
        },
        "/auth/logout": {
            "delete": {
                "security": [
                    {
                        "SessionToken": []
                    }
                ],
                "description": "Log out the current user and clear session",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Authentication"
                ],
                "summary": "User logout",
                "responses": {
                    "204": {
                        "description": "Logout successful"
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/easyflow-oauth2-server_internal_shared_errors.APIError"
                        }
                    }
                }
            }
        },
        "/auth/register": {
            "post": {
                "description": "Create a new user account with email and password",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Authentication"
                ],
                "summary": "Register a new user",
                "parameters": [
                    {
                        "description": "User registration details",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_server_routes_auth.CreateUserRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "User successfully created",
                        "schema": {
                            "$ref": "#/definitions/internal_server_routes_auth.CreateUserResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request payload",
                        "schema": {
                            "$ref": "#/definitions/easyflow-oauth2-server_internal_shared_errors.APIError"
                        }
                    },
                    "409": {
                        "description": "Email already exists",
                        "schema": {
                            "$ref": "#/definitions/easyflow-oauth2-server_internal_shared_errors.APIError"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/easyflow-oauth2-server_internal_shared_errors.APIError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "easyflow-oauth2-server_internal_shared_errors.APIError": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "Code represents the HTTP status code",
                    "type": "integer",
                    "example": 400
                },
                "details": {
                    "description": "Details contains additional error information (optional)"
                },
                "error": {
                    "description": "Error represents a predefined error code from the enum package",
                    "allOf": [
                        {
                            "$ref": "#/definitions/easyflow-oauth2-server_internal_shared_errors.ErrorCode"
                        }
                    ],
                    "example": "INVALID_REQUEST_BODY"
                }
            }
        },
        "easyflow-oauth2-server_internal_shared_errors.ErrorCode": {
            "type": "string",
            "enum": [
                "UNAUTHORIZED",
                "NOT_ALLOWED",
                "NOT_FOUND",
                "ALREADY_EXISTS",
                "INTERNAL_SERVER_ERROR",
                "MISSING_SESSION_TOKEN",
                "INVALID_SESSION_TOKEN",
                "MISSING_CLIENT_ID",
                "INVALID_CLIENT_ID",
                "MISSING_CLIENT_SECRET",
                "INVALID_CLIENT_SECRET",
                "MISSING_CODE_CHALLENGE",
                "MISSING_STATE",
                "INVALID_STATE",
                "MISSING_REDIRECT_URI",
                "INVALID_REDIRECT_URI",
                "INVALID_CONTENT_TYPE",
                "INVALID_REQUEST_BODY",
                "MISSING_GRANT_TYPE",
                "INVALID_GRANT_TYPE",
                "MISSING_CODE",
                "INVALID_CODE",
                "MISSING_CODE_VERIFIER",
                "INVALID_CODE_VERIFIER",
                "MISSING_REFRESH_TOKEN",
                "INVALID_REFRESH_TOKEN"
            ],
            "x-enum-varnames": [
                "Unauthorized",
                "NotAllowed",
                "NotFound",
                "AlreadyExists",
                "InternalServerError",
                "MissingSessionToken",
                "InvalidSessionToken",
                "MissingClientID",
                "InvalidClientID",
                "MissingClientSecret",
                "InvalidClientSecret",
                "MissingCodeChallenge",
                "MissingState",
                "InvalidState",
                "MissingRedirectURI",
                "InvalidRedirectURI",
                "InvalidContentType",
                "InvalidRequestBody",
                "MissingGrantType",
                "InvalidGrantType",
                "MissingCode",
                "InvalidCode",
                "MissingCodeVerifier",
                "InvalidCodeVerifier",
                "MissingRefreshToken",
                "InvalidRefreshToken"
            ]
        },
        "internal_server_routes_auth.CreateUserRequest": {
            "type": "object",
            "required": [
                "email",
                "password"
            ],
            "properties": {
                "email": {
                    "description": "User's email address",
                    "type": "string",
                    "example": "user@example.com"
                },
                "first_name": {
                    "description": "FirstName and LastName are optional fields",
                    "type": "string",
                    "example": "John"
                },
                "last_name": {
                    "description": "User's last name (optional)",
                    "type": "string",
                    "example": "Doe"
                },
                "password": {
                    "description": "User's password (minimum 8 characters)",
                    "type": "string",
                    "minLength": 8,
                    "example": "securePassword123"
                }
            }
        },
        "internal_server_routes_auth.CreateUserResponse": {
            "type": "object",
            "properties": {
                "email": {
                    "description": "User's email address",
                    "type": "string",
                    "example": "user@example.com"
                },
                "first_name": {
                    "description": "User's first name",
                    "type": "string",
                    "example": "John"
                },
                "id": {
                    "description": "User's unique identifier",
                    "type": "string",
                    "example": "550e8400-e29b-41d4-a716-446655440000"
                },
                "last_name": {
                    "description": "User's last name",
                    "type": "string",
                    "example": "Doe"
                }
            }
        },
        "internal_server_routes_auth.LoginRequest": {
            "type": "object",
            "required": [
                "email",
                "password"
            ],
            "properties": {
                "email": {
                    "description": "User's email address",
                    "type": "string",
                    "example": "user@example.com"
                },
                "password": {
                    "description": "User's password",
                    "type": "string",
                    "example": "securePassword123"
                }
            }
        },
        "internal_server_routes_auth.LoginResponse": {
            "type": "object",
            "properties": {
                "expiresIn": {
                    "description": "Token expiration time in seconds",
                    "type": "integer",
                    "example": 3600
                },
                "session_token": {
                    "description": "JWT session token",
                    "type": "string",
                    "example": "eyJhbGciOiJFZERTQSIsInR5cCI6IkpXVCJ9..."
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BasicAuth": {
            "type": "basic"
        },
        "SessionToken": {
            "description": "Session token for authenticated users",
            "type": "apiKey",
            "name": "session_token",
            "in": "cookie"
        }
    }
}